Basic working QAM modulation

+++ Software/Signalgenerator/Applications/Stream.hpp
@@ -5,7 +5,7 @@
 namespace Stream {
 
 void Init();
-uint8_t *GetData(uint16_t maxlen, uint16_t *len);
+uint16_t *GetData(uint16_t maxlen, uint16_t *len);
 uint8_t LoadToFPGA(uint32_t maxTime);
 
 }
